pyproject.toml
[tool.poetry]
name = "personsite"
version = "0.1.0"
description = ""
authors = ["Artemowkin <artyom.groshev2017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
Pillow = "^8.3.1"
django-cors-headers = "^3.7.0"
dj-rest-auth = "^2.1.9"
django-allauth = "^0.45.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
